% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bsvygmm.R
\name{Bsvygmm}
\alias{Bsvygmm}
\title{Bayesian growth mixture model for complex survey data.}
\usage{
Bsvygmm(
  K,
  W,
  B,
  ADJ,
  Y,
  Vr,
  Vf,
  subjectID,
  clusterIDObs,
  stratumIDObs,
  clusterIDSub,
  stratumIDSub,
  spline,
  modelType,
  priors,
  hierVar,
  inits,
  n.samples,
  burn,
  monitor,
  update,
  writeSamples
)
}
\arguments{
\item{K}{A scalar for the number of assumed unobserved, latent classes.}

\item{W}{A design matrix of an intercept and covariates for the latent class membership model, with column names to be used in MCMC storage.}

\item{B}{A design matrix of basis functions for the B-spline in the latent class membership model.}

\item{ADJ}{A binary symmetric adjacency matrix. Equal to \code{NULL} if spatial correlations are not included in the latent class membership model.}

\item{Y}{Longitudinal measurements.}

\item{Vr}{A design matrix for the subject-level random effects in the longitudinal outcome model. \code{Vr} is a subset or equal to \code{Vf}.}

\item{Vf}{A design matrix for fixed effects in the longitudinal outcome model, with column names to be used in MCMC storage.}

\item{subjectID}{Subject identifier for each observation.}

\item{clusterIDObs}{Area segment or cluster identifier for each observation.}

\item{stratumIDObs}{Stratum identifier for each observation.}

\item{clusterIDSub}{Area segment or cluster identifier for each subject.}

\item{stratumIDSub}{Stratum identifier for each subject.}

\item{spline}{Logical. If \code{TRUE}, a B-spline is included for the basis functions in argument \code{B} for the latent class membership model.}

\item{modelType}{A string. For the latent class memberhsip model, if \code{modelType} is \code{Unstr}, then the latent class membership model includes only cluster level random effects that account for within-cluster correlations. If \code{Str}, then the latent class membership model includes only cluster level random effects that account for spatial correlations among clusters. If \code{Both}, then both types of random effects are included.}

\item{priors}{A nested list of prior parameters. All regression coefficients are assigned a normal prior distribution. All scalar variance terms are assigned an inverse gamma prior distribution or a uniform prior distribution, according to user choice. The variance-covariance matrix for the subject level random effects is assigned an inverse-Wishart distribution. If any parameters are not needed in the model, the element in the list can be set to NULL. The prior distributions are not allowed to vary by latent class. Details are provided in the \code{Details} section.}

\item{hierVar}{A list with two string elements to indicate, in order, the type of prior distribution for the hierarchical variances for the random effects and for the observation level data variance. Each element may equal either \code{Unif} or \code{IG} for a uniform prior distribution or inverse gamma prior distribution, respectively.}

\item{inits}{A list of initial values. If any parameters are not needed in the model, the element in the list can be set to NULL. The initial values are allowed to vary by latent class. Details are provided in the \code{Details} section.}

\item{n.samples}{Number of MCMC iterations.}

\item{burn}{Number of MCMC samples to discard.}

\item{monitor}{Logical. If \code{TRUE}, then plots of the current iteration for the first four \code{beta} will appear, after burn-in.}

\item{update}{A scalar for the interval at which to print the current iteration. If monitor is \code{TRUE}, then trace plots will also be updated at this interval.}

\item{writeSamples}{Logical. If \code{TRUE}, then MCMC samples of all of the random effects, draws from the posterior predictive distribution, and the discrepancy measure used to evaluate overall model fit are written to text files, after burn-in. The text files for each of the random effects are accompanied by corresponding text files with column name information.}
}
\value{
Model summaries are printed to the console, including posterior means and 95\% credible intervals, posterior latent class assignment, and model comparison statistics. A label switching diagnostic using Stephen's method from the \code{label.switching} package in \code{R} is printed.

A list of matrices of stored MCMC samples (\code{n.samples-burn}) for unknown parameters, including: \enumerate{
\item \code{store_delta}: Regression coefficients from the latent class membership model.
\item \code{store_gamma2}: Variance of the stratum-level random effects in the latent class membership model.
\item \code{store_xi2}: Variance of the cluster-level spatial random effects in the latent class membership model.
\item \code{store_tau2}: Variance of the cluster-level independent random effects in the latent class membership model.
\item \code{store_alpha}: Regression coefficients for the B-spline in the latent class membership model.
\item \code{store_beta}: Regression coefficients for the longitudinal outcomes model.
\item \code{store_psi2}: Variance of the stratum-level random effects in the longitudinal outcomes model.
\item \code{store_omega2}: Variance of the cluster-level random effects in the longitudinal outcomes model.
\item \code{store_phi}: Variance-covariance of the subject-level random effects in the longitudinal outcomes model.
\item \code{store_sigma2}: Observation-level variance in the longitudinal outcomes model.
\item \code{store_pi}: Posterior probabilities of latent class assignment.
}

If \code{writeSamples = TRUE}, then samples for the random effects, draws from the posterior predictive distribution, and discrepancy measure are written to separate text files after burn-in. Corresponding text files with the column names are also written.

For the latent class membership model, the stratum-level random effects are stored in \code{store_lambda.txt}; the spatial cluster-level random effects are stored in \code{store_nu.txt}; and the independent cluster-level random effects the subject-level random effects are stored in \code{store_u.txt}.

For the longitudinal outcomes model, the stratum-level random effects are stored in \code{store_zeta.txt}; the cluster-level random effects are stored in \code{store_rho.txt}; the subject-level random effects are stored in \code{store_b.txt}.

Draws from the posterior predictive distribution are stored in \code{store_Ydraw.txt}. This file is of dimension \code{n.samples - burn} by \code{N}, where \code{N} is the number of observations (i.e., \code{length(Y)}).

Samples of a measure of discrepancy are stored in \code{store_T.txt}. This file is of dimension \code{n.samples - burn} by \code{2}, where the first column is the discrepancy measure computed using the observed data, and the second column is the discrepancy measure using the replicated data. See \code{?get_discrepancy_plot}.
}
\description{
A Bayesian Growth Mixture Model (GMM) for complex survey data is fit using a Gibbs sampler with only closed-form full conditionals.

The model was constructed for data with longitudinal measurements nested in subjects, subjects nested in area segments (also called clusters), and area segments nested in strata.

A multinomial probit model is use to predict each subject's latent class membership. The multinomial probit model uses subjects as the unit of analysis and includes area segment and stratum specific random effects to model unobserved variations at the area segment and stratum-levels. For the area segment random effects, the model includes options for modeling either or both correlations among subjects within the same area segment and spatial correlations among neighboring area segments. Correlations among subjects within the same area segment are modeled using independent random effects assumed to follow a normal distribution. Spatial correlations among neighboring area segments are modeled with an intrinsic conditional autoregessive prior distribution. An option is included to use B-splines for flexibly modeling the relationship between one of the variables (e.g., the size variable used in probability proportional to size sampling) and probability of belonging to a latent class.

In the longitudinal outcomes model, latent class-specific longitudinal trajectories are estimated. Area segment and stratum specific independent random effects are modeled using latent-class specific variance parameters. The random effects are assumed to follow a normal distribution. The observation-level variances are also latent class-specific.
}
\details{
Details are given on the specification of \code{priors} and \code{inits}.

\describe{
\item{A.	The nested list \code{priors} must contain the information on the prior distributions in the order as given below.}{
\enumerate{
\item Multivariate normal prior distribution on the regression coefficients in the latent class membership model. The list entry is of the form, for example, \code{list(rep(0, s), diag(10, s))}, where \code{s} is the number of regression coefficients. \code{rep(0, s)} assigns a prior mean of 0 to all regression coefficients. A diagonal variance-covariance matrix with variance 10 is assigned.

\item	An inverse gamma distribution on the variance of the independent random effects in the latent class membership model at the stratum level. The form is, for example, \code{list(0.1, 0.1)}, where the first element is the shape parameter and the second parameter is the scale parameter. Set this prior equal to \code{NULL} if \code{hierVar[[1]] = Unif}, in which case a uniform prior distribution is used.

\item An inverse gamma distribution on the variance of the spatial random effects in the latent class membership model at the cluster level. The form is the same as in 2. Set this prior equal to \code{NULL} if \code{hierVar[[1]] = Unif}, in which case a uniform prior distribution is used. If spatial random effects are not desired, then use \code{list(NULL)}.

\item An inverse gamma distribution on the variance of the independent random effects in the latent class membership model at the cluster level. The form is the same as in 2. Set this prior equal to \code{NULL} if \code{hierVar[[1]] = Unif}, in which case a uniform prior distribution is used. If independent random effects are not desired, then use \code{list(NULL)}.

\item If \code{spline = TRUE}, a multivariate normal prior distribution on the regression coefficients for the B-splines in the latent class membership model. The prior mean is fixed to 0. The user must specify the prior variance-covariance. The list entry is of the form, for example, \code{list(diag(10, r))}, where \code{r} is the number of B-spline coefficients. If \code{spline = FALSE}, then use \code{list(NULL)}.

\item Multivariate normal prior distribution on the regression coefficients in the longitudinal outcomes model. The specification is the same as in 1.

\item An inverse gamma distribution on the variance of the independent random effects in the longitudinal outcomes model at the stratum level. The specification is the same as in 2.

\item An inverse gamma distribution on the variance of the independent random effects in the longitudinal outcomes model at the cluster level. The specification is the same as in 4, except that cluster level random effects are always included in the longitudinal outcomes model.

\item Inverse-Wishart prior distribution on the variance-covariance matrix of the subject level random effects in the longitudinal outcomes model. The form is, for example, \code{list(q + 2, diag(0.25, q))}, where \code{q} is the number of random effects. The first element is the prior degrees of freedom, and the second element is the prior scale matrix of dimension \code{q} by \code{q}.

\item An inverse gamma distribution on the variance of the observation-level error in the longitudinal outcomes model. The specification is the same as in 2, except that this prior is equal to \code{NULL} if \code{hierVar[[2]] = Unif}, in which case a uniform prior distribution is used.}

An example of the completed list object for specification of the prior distributions is: \code{priors <- list(list(rep(0, s), diag(1, s)), list(.1, .1), list(.1, .1), list(.1, .1), list(diag(10, r)), list(rep(0, p), diag(10, p)), list(.1, .1), list(.1, .1), list((q + 2), diag(0.25, q)), list(.1, .1))}
}
\item{B.	The list \code{inits} must contain the information on initial values for each of the parameters, in the same order as in \code{priors}.}{
\enumerate{
\item Initial values for the regression coefficients in the latent class membership model. The list entry is of the form, for example, \code{matrix(rep(0, s * (K - 1)), nrow = s, ncol = (K - 1))}, where \code{s} is the number of regression coefficients, \code{K} is the assumed number of latent classes.

\item Initial values for the variances of the independent random effects in the latent class membership model at the stratum level. The list entry is of the form, for example, \code{rep(0.2, K - 1)}.

\item Initial values for the spatial variances. The specification is the same as in 2. If spatial random effects are not desired, then use \code{list(NULL)}.

\item Initial values for the variances of the independent random effects in the latent class membership model at the cluster level. The specification is the same as in 2. If independent random effects are not desired, then use \code{list(NULL)}.

\item If \code{spline = TRUE}, initial values for the regression coefficients for the B-splines in the latent class membership model. The form is the same as 1. For example, enter \code{matrix(0, nrow = r, ncol = (K - 1))}. If \code{spline = FALSE}, then use \code{list(NULL)}.

\item Initial values for the regression coefficients in the longitudinal outcomes model. The object is a \code{p} by \code{K} matrix, where \code{p} is the number of regression coefficients and \code{K} is the number of latent classes. An example is \code{matrix(rnorm(p*K), nrow = p, ncol = K)}.

\item Initial values for the variances of the independent random effects in the longitudinal outcomes model at the stratum level. The object is a \code{K} length vector. An example is \code{rep(0.1, K)}.

\item Initial values for the variances of the independent random effects in the longitudinal outcomes model at the cluster level. The form is the same as in 7.

\item Initial values for the variance-covariance matrix of the subject level random effects in the longitudinal outcomes model. The object is a \code{q} by \code{q} by \code{K} array, where \code{q} is the number of random effects and \code{K} is the number of assumed latent class. An example is \code{array(diag(0.5, q), dim = c(q, q, K))}.

\item Initial values for the variances of the observation-level errors in the longitudinal outcomes model. The form is the same as in 7.}

An example of the completed list object for specification of the initial values is:
\code{inits <- list(matrix(rep(0, s * (K - 1)), nrow = s, ncol
= (K - 1)), rep(0.2, K - 1), rep(0.2, K - 1), rep(0.2, K - 1), matrix(0, nrow = r, ncol = (K - 1)), matrix(rnorm(p * K), nrow = p, ncol = K), rep(0.1, K), rep(0.1, K), array(diag(0.5, q), dim = c(q, q, K)), rep(1, K))}
}
}
}
\examples{

#-------- Load data from Bsvygmm
#Data
data(data)
#Adjacency matrix
data(ADJ)
modelType <- "Both"
#Number of latent classes
K <- 2
#Outcome
Y <- data$Y

#Design matrices for latent class membership model, random effects, and fixed effects
dats <- aggregate(data[ , c("subjectID", "clusterID", "stratumID", "x1")], by = list(data$subjectID), FUN = tail, n = 1)
W <- model.matrix( ~ x1, data = dats)
s <- ncol(W)
#In this analysis, model time using dummies
timedf <- data.frame(time = factor(data$time))
Vr <- data.matrix(dummy(timedf, int = TRUE))
colnames(Vr) <- c("time1", "time2", "time3")
q <- ncol(Vr)
Vf <- Vr
p <- ncol(Vf)
#Identifiers at the subject and observation level
clusterIDSub <- dats$clusterID
stratumIDSub <- dats$stratumID
clusterIDObs <- data$clusterID
stratumIDObs <- data$stratumID
subjectID <- data$subjectID

#Prior for hierarchical and observation level variance
hierVar <- list("IG", "IG")

#Priors
priors <- list(list(rep(0, s), diag(1, s)), list(.1, .1), list(2, 1),
     list(.1, .1), list(NULL), list(rep(0, p), diag(10, p)),
     list(.1, .1), list(.1, .1), list((q + 2), diag(0.25, q)), list(.1, .1))

#Initial values
inits <- list(matrix(rep(0, s * (K - 1)), nrow = s, ncol = (K - 1)), rep(0.2, K - 1),
     rep(0.2, K - 1), rep(0.2, K - 1), NULL, matrix(rnorm(p * K), nrow = p, ncol = K),
     rep(0.1, K), rep(0.1, K), array(diag(0.5, q),
     dim = c(q, q, K)), rep(1, K))

#MCMC algorithm
n.samples <- 2000
burn <- 1000
update <- 10
monitor <- TRUE
res <- Bsvygmm(K = K, W = W, B = NULL, ADJ = ADJ, Y = Y,
     Vr = Vr, Vf = Vf, subjectID = subjectID,
     clusterIDObs = clusterIDObs, stratumIDObs = stratumIDObs,
     clusterIDSub = clusterIDSub, stratumIDSub = stratumIDSub,
     spline = FALSE, modelType = modelType, priors = priors,
     hierVar = hierVar, inits = inits, n.samples = n.samples,
     burn = burn, monitor = monitor, update = update,
     writeSamples = TRUE)

#Get Bayesian posterior predictive p-value, see ?get_discrepancy_plot
#Set working directory to where discrepancy samples are written
store_T <- read.table("store_T.txt", header = FALSE, sep = ",")
get_discrepancy_plot(store_T)
}
