% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_discrepancy_plot.R
\name{get_discrepancy_plot}
\alias{get_discrepancy_plot}
\title{Get discrepancy plot using observed and replicated data.}
\usage{
get_discrepancy_plot(discrepancySamples)
}
\arguments{
\item{discrepancySamples}{Store discrepancy samples from \code{Bsvygmm} using \code{writeSamples = TRUE}.}
}
\value{
A scatter plot of the discrepancy measure calculated using the observed data versus the replicated data. The plot is annotated with the posterior predictive p-value.
}
\description{
Overall adequacy of the fitted model can be examine using Bayesian posterior predictive p-values. At each MCMC iteration, a discrepancy measure is computed using the replicated and observed data. The Bayesian predictive p-value denotes the probability that the discrepancy measure under the replicated data is greater than that under the observed data. A p-value near 0.5 indicates adequate model fit, while a p-value outside the range of 0.05 and 0.95 is considered to suggest a lack of model fit. The discrepancy measure used in this function is a weighted mean squared error, computed as in:

Neelon, B., Swamy, G.K., Burgette, L.F. and Miranda, M.L. (2011) A Bayesian Growth Mixture Model to Examine Maternal Hypertension and Birth Outcomes. Statistics in Medicine, 30, 2721-2735.
}
