% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ESS.R
\name{get_ESS}
\alias{get_ESS}
\title{Get the effective sample size.}
\usage{
get_ESS(sigma2, Phi, Omega, Psi, priorPik, Y, Vr, subjectID)
}
\arguments{
\item{sigma2}{Current iteration of observation level variance.}

\item{Phi}{Current iteration of between subject variance-covariance.}

\item{Omega}{Current iteration of between area segment level variance. Equals \code{NULL} if not desired.}

\item{Psi}{Current iteration of between stratum level variance.  Equals \code{NULL} if not desired.}

\item{priorPik}{Current iteration of prior probability of class membership.}

\item{Y}{Longitudinal outcomes.}

\item{Vr}{Design matrix for random effects.}

\item{subjectID}{Subject ID for each observation.}
}
\value{
Effective sample size at the current iteration.
}
\description{
The reference for the approach is in Jones, R. H. (2011). Bayesian information criterion for longitudinal and clustered data. Statistics in Medicine, 30(25), 3050-3056.
}
