% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log_sum_exponential.R
\name{get_log_sum_exponential}
\alias{get_log_sum_exponential}
\title{Compute the log sum of exponentials.}
\usage{
get_log_sum_exponential(x)
}
\arguments{
\item{x}{Data.}
}
\value{
The log sum of x.
}
\description{
The reference for the approach is in the Stan Manual. The computation is for the log sum.
}
