% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observed_llik_sub.R
\name{get_observed_llik_sub}
\alias{get_observed_llik_sub}
\title{Get the observed data likelihood after integrating out the random effects.}
\usage{
get_observed_llik_sub(
  priorPik,
  beta,
  sigma2,
  Phi,
  Omega,
  Psi,
  Y,
  Vf,
  Vr,
  subjectID
)
}
\arguments{
\item{priorPik}{Current iteration of prior probability of class membership.}

\item{beta}{Current iteration of fixed effects.}

\item{sigma2}{Current iteration of observation level variance.}

\item{Phi}{Current iteration of between subject variance-covariance.}

\item{Omega}{Current iteration of between area segment level variance. Equals \code{NULL} if not desired.}

\item{Psi}{Current iteration of between stratum level variance. Equals \code{NULL} if not desired.}

\item{Y}{Longitudinal outcomes.}

\item{Vf}{Design matrix for fixed effects.}

\item{Vr}{Design matrix for random effects.}

\item{subjectID}{Subject ID for each observation.}
}
\value{
Observed data likelihood contribution of each of \code{n} subjects at the current iteration.
}
\description{
The reference for the approach is in Celeux, G., Forbes, F., Robert, C. P., & Titterington, D. M. (2006). Deviance Information Criteria for Missing Data Models. Bayesian Analysis, 1(4), 651-674.
}
