% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observed_posterior.R
\name{get_observed_posterior}
\alias{get_observed_posterior}
\title{Get the log observed data posterior.}
\usage{
get_observed_posterior(
  llik,
  beta,
  Phi,
  sigma2,
  Omega,
  Psi,
  delta,
  alpha,
  u,
  deltaStratum,
  nu,
  tau2,
  gamma2,
  xi2,
  priorPik,
  ADJ,
  priors,
  hierVar,
  modelType,
  spline
)
}
\arguments{
\item{llik}{Current iteration of the log of the observed data likelihood.}

\item{beta}{Current iteration of fixed effects.}

\item{Phi}{Current iteration of between subject variance-covariance.}

\item{sigma2}{Current iteration of observation level variance.}

\item{Omega}{Current iteration of between area segment level variance.}

\item{Psi}{Current iteration of between stratum level variance.}

\item{delta}{Current iteration of regression coefficients in latent class membership model.}

\item{alpha}{Current iteration of spline regression coefficients in latent class membership model.}

\item{u}{Current iteration of area segment unstructured random effects in latent class membership model.}

\item{deltaStratum}{Current iteration of stratum random effects in latent class membership model.}

\item{nu}{Current iteration of area segment structured random effects in latent class membership model.}

\item{tau2}{Current iteration of variance of area segment unstructured random effects in latent class membership model.}

\item{gamma2}{Current iteration of variance of stratum random effects in latent class membership model.}

\item{xi2}{Current iteration of variance of area segment structured random effects in latent class membership model.}

\item{priorPik}{Current iteration of prior probability of class membership.}

\item{ADJ}{Adjacency matrix used in computing prior mean of area segment structured random effects.}

\item{priors}{List of prior distributions from model fitting.}

\item{hierVar}{List of priors for variance at observation level and hierarchical variances.}

\item{modelType}{String for model type.}

\item{spline}{Logical. If \code{TRUE}, a spline is included in the latent class membership model.}
}
\value{
Observed data likelihood at the current iteration.
}
\description{
The reference for the approach is in Fruhwirth-Schnatter, S. and Pyne, S. (2010) Bayesian Inference for Finite Mixtures of Univariate and Multivariate Skew-Normal and Skew-t Distributions. Biostatistics, 11, 317-336. The observed data posterior uses the observed data likelihood after integrating out the random effects.
}
