% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_mode.R
\name{get_posterior_mode}
\alias{get_posterior_mode}
\title{Get the index for the mode of the density.}
\usage{
get_posterior_mode(samples)
}
\arguments{
\item{samples}{Samples.}
}
\value{
An index for the mode.
}
\description{
The reference for the approach is in Fruhwirth-Schnatter, S., & Pyne, S. (2010). Bayesian Inference for Finite Mixtures of Univariate and Multivariate Skew-Normal and Skew-t Distributions. Biostatistics, 11(2), 317-336.
}
