% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\name{simdat}
\alias{simdat}
\title{Simulate data for growth mixture model for complex survey data.}
\usage{
simdat(
  K,
  delta,
  gamma2,
  tau2,
  xi2,
  beta,
  Phi,
  Omega,
  Psi,
  sigma2,
  W,
  ADJ,
  Vf,
  Vr,
  subjectID,
  clusterIDObs,
  stratumIDObs,
  clusterIDSub,
  stratumIDSub,
  modelType
)
}
\arguments{
\item{K}{Number of latent classes.}

\item{delta}{A \code{K-1} column matrix of regression coefficients for the latent class membership model.}

\item{gamma2}{A \code{K-1} vector of stratum level variances for the latent class membership model.}

\item{tau2}{A \code{K-1} vector of area segment level variances for the latent class membership model.}

\item{xi2}{A \code{K-1} vector of area segment level spatial variances for the latent class membership model.}

\item{beta}{A \code{K} column matrix of regression coefficients associated with \code{Vf}.}

\item{Phi}{Variance-covariance of the subject specific effects.}

\item{Omega}{Variance at the area segment or cluster level.}

\item{Psi}{Varianec at the stratum level.}

\item{sigma2}{Observation level variance.}

\item{W}{A design matrix of an intercept and covariates for the latent class membership model.}

\item{ADJ}{A binary symmetric adjacency matrix. Equal to \code{NULL} if spatial correlations are not included in the latent class membership model.}

\item{Vf}{A design matrix for fixed effects in the longitudinal outcome model.}

\item{Vr}{A design matrix for the subject-level random effects in the longitudinal outcome model. \code{Vr} is a subset or equal to \code{Vf}.}

\item{subjectID}{Subject identifier for each observation.}

\item{clusterIDObs}{Area segment or cluster identifier for each observation.}

\item{stratumIDObs}{Stratum identifier for each observation.}

\item{clusterIDSub}{Area segment or cluster identifier for each subject.}

\item{stratumIDSub}{Stratum identifier for each subject.}

\item{modelType}{A string. For the latent class memberhsip model, if \code{modelType} is \code{Unstr}, then the latent class membership model includes only cluster level random effects that account for within-cluster correlations. If \code{Str}, then the latent class membership model includes only cluster level random effects that account for spatial correlations among clusters. If \code{Both}, then both types of random effects are included.}
}
\value{
A list with simulated longitudinal otucomes and latent class memberships.
}
\description{
Data are simulated according to the growth mixture model for complex survey data.
}
