% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_alpha_BSpline_MNP.R
\name{update_alpha_BSpline_MNP}
\alias{update_alpha_BSpline_MNP}
\title{Update regression coefficients \code{alpha} for the B-spline in the multinomial probit model.}
\usage{
update_alpha_BSpline_MNP(
  Z,
  delta,
  deltaStratum,
  u,
  nu,
  W,
  B,
  prior.Sigma,
  clusterIDSub,
  stratumIDSub
)
}
\arguments{
\item{Z}{A \code{K-1} column matrix of latent variable \code{Z}.}

\item{delta}{A \code{K-1} column matrix of regression coefficients in the multinomial probit model with \code{k=1} fixed to zero.}

\item{deltaStratum}{A \code{K-1} column matrix of stratum specific effects.}

\item{u}{A \code{K-1} column matrix of area segment specific effects for modeling correlations among subjects in the same area segment. Equals \code{NULL} if not desired.}

\item{nu}{A \code{K-1} column matrix of spatial random effects. Equals \code{NULL} if not desired.}

\item{W}{An \code{s} column design matrix of covariates, including a column of one's for intercept.}

\item{B}{An \code{R} column design matrix of basis functions for the B-spline.}

\item{prior.Sigma}{A covariance matrix supplied through \code{priors}.}

\item{clusterIDSub}{Area segment or cluster identifier for each subject.}

\item{stratumIDSub}{Stratum identifier for each subject.}
}
\value{
A \code{K-1} column matrix of regression coefficients for the B-spline with \code{k=1} fixed to zero.
}
\description{
This uses a normal prior distribution to update the regression coefficients for the B-spline in the multinomial probit model. The mean of the prior distribution is fixed to 0. The regression coefficients for \code{k=1} are fixed to zero.
}
