% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_betak_Marginal.R
\name{update_betak_Marginal}
\alias{update_betak_Marginal}
\title{Update the regression coefficients in the longitudinal outcomes model.}
\usage{
update_betak_Marginal(
  C,
  sigma2,
  Phi,
  Omega,
  Psi,
  prior.mu,
  prior.Sigma,
  Y,
  Vf,
  Vr,
  subjectID
)
}
\arguments{
\item{C}{Latent class assignments for each subject.}

\item{sigma2}{Observation level variance.}

\item{Phi}{Variance-covariance of the subject specific effects.}

\item{Omega}{Variance at the area segment or cluster level.}

\item{Psi}{Varianec at the stratum level.}

\item{prior.mu}{Prior mean.}

\item{prior.Sigma}{Prior variance.}

\item{Y}{Longitudinal measurements.}

\item{Vf}{A design matrix for fixed effects in the longitudinal outcome model.}

\item{Vr}{A design matrix for the subject-level random effects in the longitudinal outcome model. \code{Vr} is a subset or equal to \code{Vf}.}

\item{subjectID}{Subject identifier for each observation.}
}
\value{
A \code{K} column matrix of latent class specific regression coefficients.
}
\description{
A normal prior is used to update the regression coefficients in the longitudinal outcomes model. A partially marginalized sampler is used in which \code{zeta}, \code{rho}, and \code{b} are integrated out.
}
