% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_C_Marginal.R
\name{update_C_Marginal}
\alias{update_C_Marginal}
\title{Update latent class assignments.}
\usage{
update_C_Marginal(K, Z, beta, Phi, Omega, Psi, sigma2, Y, Vf, Vr, subjectID)
}
\arguments{
\item{K}{Number of latent classes.}

\item{Z}{Latent draws from the latent class membership model.}

\item{beta}{A \code{K} column matrix of regression coefficients associated with \code{Vf}.}

\item{Phi}{Variance-covariance of the subject specific effects.}

\item{Omega}{Variance at the area segment or cluster level.}

\item{Psi}{Varianec at the stratum level.}

\item{sigma2}{Observation level variance.}

\item{Y}{Longitudinal measurements.}

\item{Vf}{A design matrix for fixed effects in the longitudinal outcome model.}

\item{Vr}{A design matrix for the subject-level random effects in the longitudinal outcome model. \code{Vr} is a subset or equal to \code{Vf}.}

\item{subjectID}{Subject identifier for each observation.}
}
\value{
A list containing a vector of latent class assignments and a \code{K} column matrix of posterior probabilities of belonging to each class.
}
\description{
To update the latent class assignments, a partially marginalized sampler is used in which \code{zeta}, \code{rho}, and \code{b} are integrated out.
}
