% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_delta_BinaryP.R
\name{update_delta_BinaryP}
\alias{update_delta_BinaryP}
\title{Update regression coefficients \code{delta} in the probit model.}
\usage{
update_delta_BinaryP(
  Z,
  delta,
  alpha,
  deltaStratum,
  u,
  nu,
  W,
  B,
  prior.mu,
  prior.Sigma,
  clusterIDSub,
  stratumIDSub,
  spline
)
}
\arguments{
\item{Z}{A \code{K-1} column matrix of latent variable \code{Z}.}

\item{delta}{A \code{K-1} column matrix of regression coefficients in the multinomial probit model with \code{k=1} fixed to zero.}

\item{alpha}{A \code{K-1} column matrix of regression coefficients for the B-spline.}

\item{deltaStratum}{A \code{K-1} column matrix of stratum specific effects. Equals \code{NULL} if not desired.}

\item{u}{A \code{K-1} column matrix of area segment specific effects for modeling correlations among subjects in the same area segment. Equals \code{NULL} if not desired.}

\item{nu}{A \code{K-1} column matrix of spatial random effects. Equals \code{NULL} if not desired.}

\item{W}{An \code{s} column design matrix of covariates, including a column of one's for intercept.}

\item{B}{An \code{R} column design matrix of basis functions for the B-spline.}

\item{prior.mu}{A vector of prior means supplied through \code{priors}.}

\item{prior.Sigma}{A covariance matrix supplied through \code{priors}.}

\item{clusterIDSub}{Area segment or cluster identifier for each subject.}

\item{stratumIDSub}{Stratum identifier for each subject.}

\item{spline}{Logical. If \code{TRUE}, a B-spline is included for the basis functions in argument \code{B} for the latent class membership model.}
}
\value{
A \code{K-1} column matrix of regression coefficients in the multinomial probit model with \code{k=1} fixed to zero.
}
\description{
This uses a normal prior distribution to update the regression coefficients in the probit model. The regression coefficients for \code{k=1} are fixed to zero.
}
