% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_deltaStratum_BinaryP.R
\name{update_deltaStratum_BinaryP}
\alias{update_deltaStratum_BinaryP}
\title{Update stratum level intercepts in the probit model.}
\usage{
update_deltaStratum_BinaryP(
  Z,
  delta,
  alpha,
  u,
  nu,
  gamma2,
  W,
  B,
  clusterIDSub,
  stratumIDSub,
  spline
)
}
\arguments{
\item{Z}{\code{K-1} column matrix of latent variables Z.}

\item{delta}{\code{K-1} column matrix of latent class-specific regression coefficients.}

\item{alpha}{\code{K-1} column matrix of regression coefficients for the B-spline.}

\item{u}{A \code{K-1} column matrix of area segment specific effects for modeling correlations among subjects in the same area segment. Equals \code{NULL} if not desired.}

\item{nu}{A \code{K-1} column matrix of spatial random effects. Equals \code{NULL} if not desired.}

\item{gamma2}{A \code{K-1} vector of variances for the stratum level intercepts.}

\item{W}{An \code{s} column design matrix of covariates, including a column of one's for intercept.}

\item{B}{An \code{R} column design matrix of basis functions for the B-spline.}

\item{clusterIDSub}{Area segment or cluster identifier for each subject.}

\item{stratumIDSub}{Stratum identifier for each subject.}

\item{spline}{Logical. If \code{TRUE}, a B-spline is included for the basis functions in argument \code{B} for the latent class membership model.}
}
\value{
A \code{K - 1} matrix of stratum specific intercepts \code{delta_{sk}}.
}
\description{
This uses a normal prior distribution with mean 0 to update the stratum level intercepts in the probit model.
}
