% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_nu.R
\name{update_nu}
\alias{update_nu}
\title{Update spatial area segment (aka cluster) level effects \code{nu} in the latent class membership model.}
\usage{
update_nu(
  Z,
  delta,
  alpha,
  deltaStratum,
  u,
  nu,
  xi2,
  W,
  B,
  ADJ,
  clusterIDSub,
  stratumIDSub,
  spline
)
}
\arguments{
\item{Z}{\code{K-1} column matrix of latent variables Z.}

\item{delta}{\code{K-1} column matrix of latent class-specific regression coefficients.}

\item{alpha}{\code{K-1} column matrix of regression coefficients for the B-spline.}

\item{deltaStratum}{A \code{K-1} column matrix of stratum specific effects. Equals \code{NULL} if not desired.}

\item{u}{A \code{K-1} column matrix of area segment specific effects for modeling correlations among subjects in the same area segment. Equals \code{NULL} if not desired.}

\item{nu}{A \code{K-1} column matrix of spatial random effects.}

\item{xi2}{A \code{K-1} vector of variances for the arae segment level intercepts.}

\item{W}{An \code{s} column design matrix of covariates, including a column of one's for intercept.}

\item{B}{An \code{R} column design matrix of basis functions for the B-spline.}

\item{ADJ}{A binary symmetric adjacency matrix. Equal to \code{NULL} if spatial correlations are not included in the latent class membership model.}

\item{clusterIDSub}{Area segment or cluster identifier for each subject.}

\item{stratumIDSub}{Stratum identifier for each subject.}

\item{spline}{Logical. If \code{TRUE}, a B-spline is included for the basis functions in argument \code{B} for the latent class membership model.}
}
\value{
A \code{K - 1} matrix of spatial intercepts at the area segment level.
}
\description{
The area segment (aka cluster) level effects that model spatial correlations among neighboring area segments are updated using an intrinsic conditional autoregressive prior distribution..
}
