% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Omega.R
\name{update_Omega}
\alias{update_Omega}
\title{Update area segment (aka cluster) level variance in the longitudinal outcomes model.}
\usage{
update_Omega(C, rho, prior.scale, prior.shape, hierVar)
}
\arguments{
\item{C}{Latent class assignments for each subject.}

\item{rho}{A \code{K} column matrix of intercepts at the area segment level.}

\item{prior.scale}{A scalar for the prior scale, if an inverse gamma prior is used.}

\item{prior.shape}{A scalar for the prior shape, if an inverse gamma prior is used.}

\item{hierVar}{A string equal to \code{Unif} or \code{IG} to indicate a uniform or inverse gamma prior respectively.}
}
\value{
A \code{K} length vector of latent class specific variances.
}
\description{
An inverge gamma or uniform prior is used to update the variance of the stratum level intercepts.
}
