% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Phi.R
\name{update_Phi}
\alias{update_Phi}
\title{Update variance-covariance matrix of the subject level effects.}
\usage{
update_Phi(C, q, b, prior.df, prior.scale)
}
\arguments{
\item{C}{Latent class assignments for each subject.}

\item{q}{A scalar for the number of subject effects.}

\item{b}{A \code{q} column matrix of subject level effects.}

\item{prior.df}{A scalar for the prior degrees of freedom.}

\item{prior.scale}{A \code{q} by \code{q} prior scale matrix.}
}
\value{
Latent class specific variance-covariances.
}
\description{
An inverse Wishart prior is used to update the variance-covariance of the subject level effects.
}
