% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_priorPik.R
\name{update_priorPik}
\alias{update_priorPik}
\title{Update prior probabilities for the latent class membership model.}
\usage{
update_priorPik(Z)
}
\arguments{
\item{Z}{A \code{K-1} column matrix of latent normal variables in the multinomial probit model.}
}
\value{
An \code{K} matrix of prior probabilities.
}
\description{
This function calculates the multinomial or binary probabilities based on the assumption of independence among latent classes.
}
