% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sigma2.R
\name{update_sigma2}
\alias{update_sigma2}
\title{Update the observation level variance in the longitudinal outcomes model.}
\usage{
update_sigma2(
  C,
  zeta,
  rho,
  b,
  beta,
  Y,
  Vr,
  Vf,
  prior.shape,
  prior.scale,
  subjectID,
  clusterIDObs,
  stratumIDObs,
  hierVar
)
}
\arguments{
\item{C}{Latent class assignments for each subject.}

\item{zeta}{A \code{K} column matrix of intercepts at the stratum level.}

\item{rho}{A \code{K} column matrix of intercepts at the area segment level.}

\item{b}{A \code{q} column matrix of subject specific effects.}

\item{beta}{A \code{K} column matrix of regression coefficients associated with \code{Vf}.}

\item{Y}{Longitudinal measurements.}

\item{Vr}{A design matrix for the subject-level random effects in the longitudinal outcome model. \code{Vr} is a subset or equal to \code{Vf}.}

\item{Vf}{A design matrix for fixed effects in the longitudinal outcome model.}

\item{prior.shape}{A scalar for the prior shape, if an inverse gamma prior is used.}

\item{prior.scale}{A scalar for the prior scale, if an inverse gamma prior is used.}

\item{subjectID}{Subject identifier for each observation.}

\item{clusterIDObs}{Area segment or cluster identifier for each observation.}

\item{stratumIDObs}{Stratum identifier for each observation.}

\item{hierVar}{A string equal to \code{Unif} or \code{IG} to indicate a uniform or inverse gamma prior respectively.}
}
\value{
A \code{K} length vector of latent class specific variances.
}
\description{
An inverge gamma or uniform prior is used to update the variance at the observation level.
}
