% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tau2.R
\name{update_tau2}
\alias{update_tau2}
\title{Update variance \code{tau2} of stratum level intercepts in the latent class membership model.}
\usage{
update_tau2(u, prior.shape, prior.scale, hierVar)
}
\arguments{
\item{u}{A \code{K-1} column matrix of area segment level intercepts.}

\item{prior.shape}{A scalar for the prior shape, if an inverse gamma prior is used.}

\item{prior.scale}{A scalar for the prior scale, if an inverse gamma prior is used.}

\item{hierVar}{A string equal to \code{Unif} or \code{IG} to indicate a uniform prior distribution or inverse gamma prior distribution.}
}
\value{
A \code{K-1} vector of variances.
}
\description{
This uses an inverse gamma prior or a uniform prior to update the variance of the area segment (aka cluster) level intercepts in the latent class membership model.
}
