% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_u_MNP.R
\name{update_u_MNP}
\alias{update_u_MNP}
\title{Update area segment (aka cluster) level intercepts in the multinomial probit model.}
\usage{
update_u_MNP(
  Z,
  delta,
  alpha,
  deltaStratum,
  nu,
  tau2,
  W,
  B,
  clusterIDSub,
  stratumIDSub,
  spline
)
}
\arguments{
\item{Z}{A \code{K-1} column matrix of latent variable \code{Z}.}

\item{delta}{A \code{K-1} column matrix of regression coefficients in the multinomial probit model with \code{k=1} fixed to zero.}

\item{alpha}{A \code{K-1} column matrix of regression coefficients for the B-spline.}

\item{deltaStratum}{A \code{K-1} column matrix of stratum specific effects.}

\item{nu}{A \code{K-1} column matrix of spatial random effects. Equals \code{NULL} if not desired.}

\item{W}{An \code{s} column design matrix of covariates, including a column of one's for intercept.}

\item{B}{An \code{R} column design matrix of basis functions for the B-spline.}

\item{clusterIDSub}{Area segment or cluster identifier for each subject.}

\item{stratumIDSub}{Stratum identifier for each subject.}

\item{spline}{Logical. If \code{TRUE}, a B-spline is included for the basis functions in argument \code{B} for the latent class membership model.}
}
\value{
A \code{K - 1} matrix of area segment specific intercepts \code{u}.
}
\description{
This uses a normal prior distribution with mean 0 to update the area segment (aka cluster) level intercepts in the multinomial probit model. These intercepts model correlation among subjects in the same area segment.
}
