% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_xi2.R
\name{update_xi2}
\alias{update_xi2}
\title{Update variance \code{xi2} of area segment level spatial intercepts in the latent class membership model.}
\usage{
update_xi2(nu, ADJ, prior.shape, prior.scale, hierVar)
}
\arguments{
\item{nu}{A \code{K-1} column matrix of area segment level spatial intercepts.}

\item{ADJ}{A binary symmetric adjacency matrix. Equal to \code{NULL} if spatial correlations are not included in the latent class membership model.}

\item{prior.shape}{A scalar for the prior shape, if an inverse gamma prior is used.}

\item{prior.scale}{A scalar for the prior scale, if an inverse gamma prior is used.}

\item{hierVar}{A string equal to \code{Unif} or \code{IG} to indicate a uniform prior distribution or inverse gamma prior distribution.}
}
\value{
A \code{K-1} vector of variances.
}
\description{
This uses an inverse gamma prior or a uniform prior to update the variance of the area segment (aka cluster) level spatial intercepts in the latent class membership model.
}
