% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Z_MNP.R
\name{update_Z_MNP}
\alias{update_Z_MNP}
\title{Update latent variable \code{Z} in multinomial probit model.}
\usage{
update_Z_MNP(
  Z,
  C,
  delta,
  alpha,
  deltaStratum,
  u,
  nu,
  W,
  B,
  clusterIDSub,
  stratumIDSub,
  spline
)
}
\arguments{
\item{Z}{\code{K-1} column matrix of latent variables Z.}

\item{C}{Latent class assignments for each subject.}

\item{delta}{\code{K-1} column matrix of latent class-specific regression coefficients.}

\item{alpha}{\code{K-1} column matrix of regression coefficients for the B-spline.}

\item{deltaStratum}{A \code{K-1} column matrix of stratum specific effects.}

\item{u}{A \code{K-1} column matrix of area segment specific effects for modeling correlations among subjects in the same area segment. Equals \code{NULL} if not desired.}

\item{nu}{A \code{K-1} column matrix of spatial random effects. Equals \code{NULL} if not desired.}

\item{W}{An \code{s} column design matrix of covariates, including a column of one's for intercept.}

\item{B}{An \code{R} column design matrix of basis functions for the B-spline.}

\item{clusterIDSub}{Area segment or cluster identifier for each subject.}

\item{stratumIDSub}{Stratum identifier for each subject.}

\item{spline}{Logical. If \code{TRUE}, a B-spline is included for the basis functions in argument \code{B} for the latent class membership model.}
}
\value{
A \code{K - 1} matrix of latent variable \code{Z}.
}
\description{
The latent variable in the multinomial probit model is updated using a truncated normal distribution. The updating algorithm allows for either or both area segment intercepts for correlations among subjects in the same area segment and spatial correlations among neighboring area segments. The number of latent classes is assumed to be greater than 2.
}
