% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_zeta.R
\name{update_zeta}
\alias{update_zeta}
\title{Update the stratum level intercepts in the longitudinal model.}
\usage{
update_zeta(
  C,
  b,
  rho,
  beta,
  sigma2,
  Psi,
  Y,
  Vf,
  Vr,
  subjectID,
  clusterIDObs,
  stratumIDObs
)
}
\arguments{
\item{C}{Latent class assignments for each subject.}

\item{b}{A \code{q} column matrix of subject specific effects.}

\item{rho}{A \code{K} column matrix of area segment (aka cluster) level intercepts. Equals \code{NULL} if not desired.}

\item{beta}{A \code{K} column matrix of regression coefficients associated with \code{Vf}.}

\item{sigma2}{A \code{K} length vector of observation level variances.}

\item{Psi}{A \code{K} length vector of stratum level variances.}

\item{Y}{Longitudinal measurements.}

\item{Vf}{A design matrix for fixed effects in the longitudinal outcome model.}

\item{Vr}{A design matrix for the subject-level random effects in the longitudinal outcome model. \code{Vr} is a subset or equal to \code{Vf}.}

\item{subjectID}{Subject identifier for each observation.}

\item{clusterIDObs}{Area segment or cluster identifier for each observation.}

\item{stratumIDObs}{Stratum identifier for each observation.}
}
\value{
A \code{K} column matrix of stratum level intercepts.
}
\description{
This uses a normal prior distribution with mean 0 to update the stratum level intercepts in the longitudinal outcomes model.
}
