% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_discrepancy.R
\name{get_discrepancy}
\alias{get_discrepancy}
\title{Get posterior predictive draws of discrepancy the weighted mean squared error at each iteration.}
\usage{
get_discrepancy(
  K,
  C,
  b,
  sigma2,
  beta,
  zeta,
  rho,
  Phi,
  Omega,
  Psi,
  Z,
  delta,
  alpha,
  deltaStratum,
  u,
  nu,
  gamma2,
  tau2,
  xi2,
  priorPik,
  W,
  B,
  ADJ,
  clusterIDSub,
  stratumIDSub,
  spline,
  subjectID,
  clusterIDObs,
  stratumIDObs,
  Y,
  Vf,
  Vr,
  LRModelType,
  LCClusterModelType,
  LCStratumModelType
)
}
\value{
Draws of the discrepancy measure computed at the observed data and replicated data.
}
\description{
See references for the discrepancy measure in manuscript. The algorithm used is based on redrawing the complete data. This is consistent with the function \code{get_post_pred_redraw()}.
}
