% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_post_pred.R
\name{get_postPred}
\alias{get_postPred}
\title{Get posterior predictive draws at each iteration.}
\usage{
get_postPred(
  K,
  beta,
  sigma2,
  priorPik,
  b,
  rho,
  zeta,
  Vr,
  Vf,
  subjectID,
  clusterIDObs,
  stratumIDObs
)
}
\arguments{
\item{K}{Number of latent classes.}

\item{beta}{Current iteration of fixed effects.}

\item{sigma2}{Current iteration of observation level variance.}

\item{priorPik}{Current iteration of prior probability of class membership.}

\item{b}{A \code{q} column matrix of subject-level effects.}

\item{rho}{A \code{K} column matrix of area segment level intercepts.}

\item{zeta}{A \code{K} column matrix of stratum level intercepts.}

\item{Vr}{Design matrix for random effects.}

\item{Vf}{Design matrix for fixed effects.}

\item{subjectID}{Subject identifier for each observation.}

\item{clusterIDObs}{Area segment or cluster identifier for each observation.}

\item{stratumIDObs}{Stratum identifier for each observation.}
}
\value{
Draws from the posterior predictive distribution at the current iteration.
}
\description{
The reference for the approach is from algorithm 3.7 on page 91 of Fruhwirth-Schnatter, S. (2006) Finite Mixture and Markov Switching Models. Springer Science & Business Media, New York. The latent classes are redrawn using the probabilities of latent class membership obtained from the latent class membership model.
}
